pyproject.toml
[project]
name = "recipe-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
